# WIP: use pairwise summation in `sum`