Working towards fixing gh-2855. The immediate problem is that this currently a little over 2x slower. The NumPy implementation unrolls loops manually, maybe that would help.

**Edit**: manually unrolling the loops doesn't help.

Timings:

```
In [1]: import numpy as np

In [2]: from numba import njit

In [3]: @njit
   ...: def f(A):
   ...:     return np.sum(A)
   ...:

In [4]: A = np.ones(10**6) / np.float32(255.0)

In [5]: f(A)
Out[5]: 3921.568627447501

In [6]: %timeit f(A)
```

On master:

```
779 µs ± 11.8 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
```

On this branch:

```
1.62 ms ± 13.7 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
```

Just doing it with NumPy directly:

```
In [1]: import numpy as np

In [2]: A = np.ones(10**6, dtype=np.float32) / np.float32(255.0)

In [3]: %timeit np.sum(A)
313 µs ± 14.7 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
```