@person142 2018-04-02
Ok, the slowdown can be avoided by not allocating the array `r`; doing that in the dumbest possible way gives:

```
r0 = arr[0]
r1 = arr[1]
r2 = arr[2]
r3 = arr[3]
r4 = arr[4]
r5 = arr[5]
r6 = arr[6]
r7 = arr[7]

m = n - (n % 8)
for i in range(8, m, 8):
    r0 += arr[i]
    r1 += arr[i+1]
    r2 += arr[i+2]
    r3 += arr[i+3]
    r4 += arr[i+4]
    r5 += arr[i+5]
    r6 += arr[i+6]
    r7 += arr[i+7]
res = ((r0 + r1) + (r2 + r3)) + ((r4 + r5) + (r6 + r7))
```

Timings for that are:

```
806 µs ± 8.98 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
```

Might be able to get to NumPy performance by creating an LLVM vector and emitting `fadd` instructions...