@seibert 2018-04-05
I'm impressed you were able to get this into a form Numba would accept and generate high performing code!  Drilling back to the [NumPy PR](https://github.com/numpy/numpy/pull/3685) where this algorithm was proposed, I see the comment:

> An alternative to pairwise summation is kahan summation but in order to
have a low performance penalty one must unroll and vectorize it,
while pairwise summation has the same speed without any vectorization.

Which makes me wonder if Numba (since as a JIT, it is much better situated to do vectorization for the CPU the user has) would produce better performing code with Kahan summation.  This would also avoid the need for a recursive function call, which prevents this implementation of sum() from being inlined into other functions.