@person142 2018-04-05
> generate high performing code

I have a faster version but it involves even more hacks. ;-) Things would be simplified if there were a "high-level" way to use pointers in Numba (versus falling back to IRBuilder).

> would produce better performing code with Kahan summation

I suspect that Kahan will always be slower since it requires `~4*n` flops over pairwise summation's `~n` flops

> This would also avoid the need for a recursive function call

I should be able to make this implementation non recursive. It would also avoid taking so many views, which turns out to be a significant slowdown.