@person142 2018-04-06
Realized the views can be avoided; this gets the timings to

```
708 µs ± 1.84 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
```

i.e. a little better than master. Splitting out the inner loop and forcing fast math might help more. (Need to get those inner sums vectorized...)

A non-recursive implementation will require allocating an array; not sure how that's going to compare. I can also do a quick Kahan for comparison since it's a simple algorithm.