@person142 2018-04-07
A branch implementing Kahan is here:

https://github.com/person142/numba/blob/kahan-sum/numba/targets/arraymath.py#L156

Timings are

```
3.18 ms ± 94.9 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
```

i.e. a fairly straightforward 4x increase in time vs. master because of the increase in flops.

Also looks like not using views appears to have led to a failure in parfors; not sure what's going on there.