@stuartarchibald 2018-05-10
It seems like LLVM can't automatically remove the recursion, generated asm has e.g.:
```asm
	movabsq	$_ZN8__main__16pairwise_sum$245E5ArrayIdLi1E1A7mutable7alignedExxx, %rax
	callq	*%rax
	addq	$64, %rsp
	movsd	16(%rsp), %xmm0
	movsd	%xmm0, 40(%rsp)
	movq	$0, 8(%rsp)
	subq	$8, %rsp
	leaq	16(%rsp), %rdi
	movq	%r13, %rcx
	movq	40(%rsp), %r8
	movq	32(%rsp), %r9
	pushq	%r15
	subq	%r12, %r15
	pushq	%r12
	pushq	%r15
	pushq	%r14
	pushq	%rbp
	pushq	%rbx
	pushq	168(%rsp)
	movabsq	$_ZN8__main__16pairwise_sum$245E5ArrayIdLi1E1A7mutable7alignedExxx, %rax
	callq	*%rax
```
in the main block.