@stuartarchibald 2018-05-10
It also appears to sometimes produce the wrong answer?
```python
@njit
def foo(x):
    return np.sum(x)

x = np.arange(1e4)

print(foo(x))
print(x.sum())
```
gives:
```
776760.0
49995000.0
```