@stuartarchibald 2018-05-10
Here's a loopy version of pairwise reduction for performance comparison:
```python
@njit
def array_sum_impl(arr):
    n = arr.size
    flatarr = as_strided(arr, shape=(n,), strides=arr.strides[-1:])
    base_N = 128 # the chunk length
    tile_n = n // base_N
    acc = 0
    if tile_n > 0:
        # there is at least 1 base_N length chunk
        # trivial tree reduction needs a 2**n space
        npwr2 =  int(2 ** np.ceil(np.log2(tile_n)))
        acc_vec = np.zeros(npwr2, dtype=arr.dtype)

        # accumulate the sum of adjacent base_N chunks into a vector
        for i in range(tile_n):
            # inner loop
            inner_acc = 0
            start = i * base_N
            stop = (i + 1) * base_N
            for j in range(start, stop):
                inner_acc += flatarr[j]
            acc_vec[i] = inner_acc

        # tree reduction vector sum
        jmp = 2
        while jmp <= npwr2:
            hjmp = jmp >> 1
            for x in range(0, tile_n, jmp):
                acc_vec[x] += acc_vec[x + hjmp]
            jmp  = jmp << 1

        acc += acc_vec[0]

    # clean up remaining
    rem = n % base_N
    for i in range(n - rem, n):
        acc += flatarr[i]

    return acc
```