@person142 2018-05-20
> It also appears to sometimes produce the wrong answer?

Derp, forgot to adjust the start/stop indices in the recursive call when I stopped taking views. Fixing that should hopefully make Travis green too.