@person142 2018-05-20
For the non-recursive version I'm getting

```
637 µs ± 7.7 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
```

whereas recursive is wallowing at a slow

```
768 µs ± 15 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
```

so that's a pretty easy win for doing it the non-recursive way speed-wise.