@stuartarchibald 2018-06-21
As suggested I've added the loopy version as https://github.com/numba/numba/pull/3046, thanks for all the investigative work @person142 , that was really helpful. Closing.